changed filetype from .h to .hpp
